Common class for tank and turret was added
