+++ TankGame/Source/Tankogedon/TankPawn.h
@@ -3,14 +3,11 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "GameFramework/Pawn.h"
-#include "DamageTaker.h"
+#include "MachinePawn.h"
 #include "TankPawn.generated.h"
 
-class UStaticMeshComponent;
-class ACannon;
 UCLASS()
-class TANKOGEDON_API ATankPawn : public APawn, public IDamageTaker
+class TANKOGEDON_API ATankPawn : public AMachinePawn
 {
 	GENERATED_BODY()
 
@@ -23,56 +20,14 @@ public:
 	void MoveRight(float RightValue);
 	void RotateRight(float RotateValue);
 
-	void Fire();
-	void FireSpecial();
-	void AutomaticFire();
-
-	void ChangeCannon();
-
-	void SetupCannon(TSubclassOf<ACannon> newCannonClass);
-	ACannon* GetCannon() const;
-
-	virtual void TakeDamage(FDamageData DamageData);
-
-	UFUNCTION()
-	void DamageTaken(float Value);
-
-	UFUNCTION()
-	void Die();
-
-	bool changeFlag = false;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Cannon")
-	TSubclassOf<ACannon> CannonClass;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Cannon")
-	TSubclassOf<ACannon> SecondCannonClass;
 protected:
 	virtual void BeginPlay() override;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-	UStaticMeshComponent* BodyMesh;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-	UStaticMeshComponent* TurretMesh;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-	class UBoxComponent* BoxCollision;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-	class UHealthComponent* HealthComponent;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
 	class USpringArmComponent* SpringArm;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
 	class UCameraComponent* Camera;
-
-	UPROPERTY()
-	ACannon* Cannon;	
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Cannon")
-	class UArrowComponent* CannonSetupPoint;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Movement")
 	float MoveSpeed = 100.0f;
