+++ TankGame/Source/Tankogedon/Turret.h
@@ -2,52 +2,20 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "GameFramework/Actor.h"
-#include "Cannon.h"
-#include "DamageTaker.h"
-#include "HealthComponent.h"
+#include "MachinePawn.h"
 #include "Turret.generated.h"
 
-class UStaticMeshComponet;
-class ACannon;
 UCLASS()
-class TANKOGEDON_API ATurret : public AActor, public IDamageTaker
+class TANKOGEDON_API ATurret : public AMachinePawn
 {
 	GENERATED_BODY()
 	
 public:	
 	ATurret();
 
-	UFUNCTION()
-	virtual void TakeDamage(FDamageData DamageData) override;
-
-	UFUNCTION()
-	void Die();
-
-	UFUNCTION()
-	void DamageTaked(float Value);
-
 protected:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-		UStaticMeshComponent* BodyMesh;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-		UStaticMeshComponent* TurretMesh;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-	class UArrowComponent* CannonSetupPoint;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
 	class UBoxComponent* HitCollider;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-	class UHealthComponent* HealthComponent;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Components")
-	TSubclassOf<ACannon> CannonClass;
-
-	UPROPERTY()
-	ACannon* Cannon;
 
 	UPROPERTY()
 	class APawn* PlayerPawn;
@@ -71,10 +39,8 @@ protected:
 	virtual void BeginPlay() override;
 	virtual void Destroyed() override;
 
-	void SetupCannon(TSubclassOf<ACannon> newCannonClass);
 	void Targeting();
 	void RotateToPlayer();
-	void Fire();
 	bool IsPlayerInRange();
 	bool CanFire();
 };
